Setup test, break apart entrypoint, & align GHA workflow
