+++ backend/test-resources/docker-compose.test.yml
@@ -1,0 +1,12 @@
+version: '3'
+
+services:
+  mongo-test:
+    image: mongo
+    container_name: infisical-test-mongo
+    restart: always
+    ports:
+      - 27018:27017
+    environment:
+      - MONGO_INITDB_ROOT_USERNAME=test
+      - MONGO_INITDB_ROOT_PASSWORD=test1234
