+++ backend/test-resources/env-vars.js
@@ -1,0 +1,5 @@
+/* eslint-disable no-undef */
+process.env.MONGO_URL =
+  'mongodb://test:test1234@localhost:27018/?authSource=admin';
+process.env.MONGO_USERNAME = 'test';
+process.env.MONGO_PASSWORD = 'test1234';
